Add UserInputsException and validation

+++ src/main/java/com/amalitech/SpringBootBloggingApp/service/impl/CommentServiceImpl.java
@@ -6,6 +6,8 @@ import com.amalitech.SpringBootBloggingApp.model.entity.User;
 import com.amalitech.SpringBootBloggingApp.repository.impl.CommentRepositoryImpl;
 import com.amalitech.SpringBootBloggingApp.service.CommentService;
 
+import com.amalitech.SpringBootBloggingApp.util.ValidationUtil;
+import com.amalitech.SpringBootBloggingApp.util.exceptions.UserInputsException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -23,26 +25,44 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public Comment create(Comment comment) {
+        if(!ValidationUtil.isValidComment(comment.getContent())) {
+            throw new UserInputsException("Invalid comment");
+        }
         return commentRepository.save(comment);
     }
 
     @Override
     public boolean delete(String commentId) {
+        if(!ValidationUtil.isValidObjectId(commentId)) {
+            throw new UserInputsException("Invalid comment ID");
+        }
         return commentRepository.deleteById(commentId);
     }
 
     @Override
     public List<Comment> getByPost(String postId) {
+        if(!ValidationUtil.isValidObjectId(postId)) {
+            throw new UserInputsException("Invalid post ID");
+        }
         return commentRepository.findByPostId(postId);
     }
 
     @Override
     public List<Comment> getByUser(String userId) {
+        if(!ValidationUtil.isValidObjectId(userId)) {
+            throw new UserInputsException("Invalid user ID");
+        }
         return commentRepository.findByUserId(userId);
     }
 
     @Override
     public boolean update(Comment comment) {
+        if(!ValidationUtil.isValidObjectId(comment.getId())) {
+            throw new UserInputsException("Invalid comment ID");
+        }
+        if(!ValidationUtil.isValidComment(comment.getContent())) {
+            throw new UserInputsException("Invalid comment");
+        }
         return commentRepository.update(comment);
     }
 
@@ -53,6 +73,9 @@ public class CommentServiceImpl implements CommentService {
 
      @Override
     public Comment findById(String commentId) {
+        if(!ValidationUtil.isValidObjectId(commentId)) {
+            throw new UserInputsException("Invalid comment ID");
+        }
          Optional<Comment> comment = commentRepository.findById(commentId);
         return comment.orElse(null);
     }
